+++ jetpack_vendor/automattic/jetpack-boost-speed-score/src/contracts/boost-api-client.php
@@ -1,6 +1,11 @@
 <?php
+/**
+ * Boost API Client interface.
+ *
+ * @package automattic/jetpack-boost-speed-score
+ */
 
-namespace Automattic\Jetpack_Boost\Contracts;
+namespace Automattic\Jetpack\Boost_Speed_Score\Contracts;
 
 /**
  * An interface to build Boost API client.
@@ -12,9 +17,8 @@ interface Boost_API_Client {
 	/**
 	 * Submit a request to boost API and return response.
 	 *
-	 * @param $method string Request method POST|GET|PUT|PATCH|DELETE
-	 * @param $path string Request path
-	 * @param $args mixed[] Request arguments
+	 * @param string  $path - Request path.
+	 * @param mixed[] $payload - Request arguments.
 	 * @return mixed
 	 */
 	public function post( $path, $payload = array() );
@@ -22,8 +26,8 @@ interface Boost_API_Client {
 	/**
 	 * Make a get request to boost API and return response.
 	 *
-	 * @param $path string Request path
-	 * @param $query mixed[] Query parameters
+	 * @param string  $path - Request path.
+	 * @param mixed[] $query - Query parameters.
 	 */
 	public function get( $path, $query = array() );
 }