+++ jetpack_vendor/automattic/jetpack-boost-speed-score/src/lib/class-transient.php
@@ -2,12 +2,10 @@
 /**
  * Transients for Jetpack Boost.
  *
- * @link       https://automattic.com
- * @since      1.0.0
- * @package    automattic/jetpack-boost
+ * @package automattic/jetpack-boost-speed-score
  */
 
-namespace Automattic\Jetpack_Boost\Lib;
+namespace Automattic\Jetpack\Boost_Speed_Score\Lib;
 
 /**
  * Class Transient
@@ -92,6 +90,7 @@ class Transient {
 		 */
 		$prefix_search_pattern = $wpdb->esc_like( $option_prefix ) . '%';
 
+		//phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery,WordPress.DB.DirectDatabaseQuery.NoCaching
 		$option_names = $wpdb->get_col(
 			$wpdb->prepare(
 				"
@@ -102,6 +101,7 @@ class Transient {
 				$prefix_search_pattern
 			)
 		);
+		// phpcs:enable
 
 		// Go through each option individually to ensure caches are handled properly.
 		foreach ( $option_names as $option_name ) {